refactor(scene): simplify screenshot response
