+++ src/api/scene/scene.ts
@@ -1,3 +1,5 @@
+import sharp from 'sharp';
+import { readFile, unlink } from 'fs/promises';
 import {
     SchemaOpenOptions,
     SchemaCloseResult,
@@ -16,27 +18,32 @@ import {
     TReload,
     TSaveResult,
 } from './schema';
+import {
+    SchemaScreenshotOptions,
+    SchemaScreenshotResult,
+    TScreenshotOptions,
+    TScreenshotResult,
+} from './screenshot-schema';
 import { description, param, result, title, tool } from '../decorator/decorator.js';
 import { COMMON_STATUS, CommonResultType, getCommonErrorStatus } from '../base/schema-base';
 import { Scene, TSceneTemplateType } from '../../core/scene';
 import { ComponentApi } from './component';
 import { NodeApi } from './node';
 import { PrefabApi } from './prefab';
-import { SceneScreenshotApi } from './screenshot';
-import { options } from '../../core/builder/platforms/android/i18n/en';
+
+type TScreenshotResultWithPayload = TScreenshotResult & {
+    image: TScreenshotResult['image'] & { base64: string };
+};
 
 export class SceneApi {
     public component: ComponentApi;
     public node: NodeApi;
     public prefab: PrefabApi;
-    public screenshot: SceneScreenshotApi;
 
     constructor() {
         this.component = new ComponentApi();
         this.node = new NodeApi();
         this.prefab = new PrefabApi();
-        // 场景截图工具（实例化以触发 @tool 装饰器注册）
-        this.screenshot = new SceneScreenshotApi();
     }
 
     @tool('scene-query-current')
@@ -136,7 +143,7 @@ export class SceneApi {
                 targetDirectory: options.dbURL,
                 templateType: options.templateType as TSceneTemplateType,
             });
-            
+
             return {
                 code: COMMON_STATUS.SUCCESS,
                 data: data as TCreateResult,
@@ -167,6 +174,86 @@ export class SceneApi {
                 code: COMMON_STATUS.FAIL,
                 reason: e instanceof Error ? e.message : String(e)
             };
+        }
+    }
+
+    @tool('scene-get-screenshot')
+    @title('Capture scene screenshot') // 截取当前场景画面
+    @description(
+        'Headlessly render the currently open scene from project files and return the image (base64, for multimodal input) plus scene metadata '
+        + '(scene name, file mtime, actual camera params). '
+        + 'Use it to visually reason about layout, spacing, occlusion and spatial relations. '
+        + 'The image is scaled/compressed and is a tool-chain rendered preview, not pixel-accurate; confirm exact values with structured data.',
+    ) // 以 headless 离屏渲染当前场景，返回图片 + 场景元数据，供多模态视觉分析布局/遮挡/间距
+    @result(SchemaScreenshotResult)
+    async getSceneScreenshot(
+        @param(SchemaScreenshotOptions) options: TScreenshotOptions,
+    ): Promise<CommonResultType<TScreenshotResultWithPayload>> {
+        let tempFilePath: string | undefined;
+        try {
+            // 场景进程在截图 RPC 内安全刷新目标，避免旧实例在重新打开前被保存。
+            const shot = await Scene.Screenshot.capture({
+                sceneUrlOrUUID: options.sceneUrlOrUUID,
+                width: options.width,
+                height: options.height,
+                camera: options.camera,
+                viewMode: options.viewMode,
+                includeGizmos: options.includeGizmos,
+            });
+            tempFilePath = shot.filePath;
+
+            const format = options.format ?? 'jpeg';
+            const quality = options.quality ?? 80;
+            const mimeType = format === 'png' ? 'image/png' : 'image/jpeg';
+
+            // 默认保持设计分辨率；调用方显式指定 maxSize 时才等比缩小。
+            const raw = await readFile(shot.filePath);
+            let pipeline = sharp(raw);
+            if (options.maxSize != null) {
+                pipeline = pipeline.resize({
+                    width: options.maxSize,
+                    height: options.maxSize,
+                    fit: 'inside',
+                    withoutEnlargement: true,
+                });
+            }
+            pipeline = format === 'png'
+                ? pipeline.png({ quality, compressionLevel: 9 })
+                : pipeline.jpeg({ quality });
+
+            const { data: out, info: outputInfo } = await pipeline.toBuffer({ resolveWithObject: true });
+            const base64 = out.toString('base64');
+
+            return {
+                code: COMMON_STATUS.SUCCESS,
+                data: {
+                    image: { base64, mimeType },
+                    meta: {
+                        sceneUrl: shot.sceneUrl,
+                        sceneName: shot.sceneName,
+                        mtime: shot.mtime,
+                        actualCameras: shot.actualCameras,
+                        width: outputInfo.width,
+                        height: outputInfo.height,
+                    },
+                },
+            };
+        } catch (e) {
+            console.error(e);
+            return {
+                code: getCommonErrorStatus(e),
+                reason: e instanceof Error ? e.message : String(e),
+            };
+        } finally {
+            if (tempFilePath) {
+                try {
+                    await unlink(tempFilePath);
+                } catch (cleanupError: any) {
+                    if (cleanupError?.code !== 'ENOENT') {
+                        console.warn(`[Screenshot] Failed to remove temporary image ${tempFilePath}:`, cleanupError);
+                    }
+                }
+            }
         }
     }
 }
